Revert "slap: Use CacheControl on client side to respect HTTP cache headers"
This reverts commit `16ca1f2c455c710c0c202e0859a69bdfa3e4fe92`.

While `CacheControl` correctly implements HTTP caching, it does not fulfill the original intent of supporting the `state-if-error` use case. Improvements to keep running nodes while master is off, has since been implemented in a more effective and SlapOS-specific manner, making this change redundant.

Additionally, `CacheControl` introduces issues when processing `If-Modified-Since` headers, causing the cache to be preserved in unexpected ways.

As we do not plan to rely on HTTP caching for the new API, this change removes `CacheControl` to resolve the issues introduced—without sacrificing any functionality, since the cache was not meaningfully used.